requirements.txt
numpy~=1.22.2
omegaconf~=2.1.1
sentencepiece~=0.1.96
hydra-core~=1.1.0
nmslib~=2.1.1
pytorch-lightning~=1.4.1
pandas~=1.1.5
torch~=1.12.1
torchmetrics~=0.5.1
tqdm~=4.62.2
transformers~=4.21.3